util: Move `isSerializationFirstNode` to `runtime` package to resolve cyclic dependency
