+++ packages/@glimmer/util/index.ts
@@ -8,11 +8,6 @@ export { default as assert } from './lib/assert';
 export { assign, fillNulls } from './lib/object-utils';
 export { ensureGuid, initializeGuid, HasGuid } from './lib/guid';
 
-export {
-  isSerializationFirstNode,
-  SERIALIZATION_FIRST_NODE_STRING,
-} from './lib/is-serialization-first-node';
-
 export { Stack, Dict, Set, DictSet, dict } from './lib/collections';
 export {
   EMPTY_SLICE,
